+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - (Feature) (Gateway) SNI and Authz support
 - (Maintenance) Bump Examples to ArangoDB 3.12
 - (Feature) (Gateway) ArangoDB JWT Auth Integration
+- (Feature) Scheduler Handler
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries